chore(swap): add tab navigation
