+++ apps/wallet-mobile/src/features/Swap/useCases/StartSwapScreen/ListOrders/ListOrders.tsx
@@ -3,6 +3,7 @@ import {StyleSheet, View} from 'react-native'
 
 import {Boundary} from '../../../../../components'
 import {useSearchOnNavBar} from '../../../../../Search/SearchContext'
+import {COLORS} from '../../../../../theme'
 import {ButtonGroup} from '../../../common/ButtonGroup/ButtonGroup'
 import {useStrings} from '../../../common/strings'
 import {CompletedOrders, CompletedOrdersSkeleton} from './CompletedOrders'
@@ -52,5 +53,6 @@ const styles = StyleSheet.create({
     flex: 1,
     justifyContent: 'space-between',
     padding: 16,
+    backgroundColor: COLORS.WHITE,
   },
 })
