--- apps/wallet-mobile/src/features/Swap/useCases/StartSwapScreen/StartSwapScreen.stories.tsx
@@ -1,25 +1,0 @@
-import {storiesOf} from '@storybook/react-native'
-import {mockSwapManager, SwapProvider} from '@yoroi/swap'
-import React from 'react'
-
-import {SelectedWalletProvider} from '../../../../SelectedWallet'
-import {mocks as walletMocks} from '../../../../yoroi-wallets/mocks'
-import {SwapFormProvider} from '../../common/SwapFormProvider'
-import {StartSwapScreen} from './StartSwapScreen'
-
-storiesOf('Swap Start Screen', module) //
-  .add('initial state', () => {
-    return <Initial />
-  })
-
-const Initial = () => {
-  return (
-    <SelectedWalletProvider wallet={walletMocks.wallet}>
-      <SwapProvider swapManager={mockSwapManager}>
-        <SwapFormProvider>
-          <StartSwapScreen />
-        </SwapFormProvider>
-      </SwapProvider>
-    </SelectedWalletProvider>
-  )
-}
