--- apps/wallet-mobile/src/features/Swap/useCases/StartSwapScreen/index.ts
@@ -1,1 +1,0 @@
-export * from './StartSwapScreen'
