+++ apps/wallet-mobile/src/features/Swap/useCases/index.ts
@@ -1,4 +1,3 @@
 export * from './ConfirmTxScreen/ConfirmTxScreen'
-export * from './StartSwapScreen'
 export * from './StartSwapScreen/CreateOrder/EditPool/SelectPoolFromListScreen/SelectPoolFromListScreen'
 export * from './StartSwapScreen/CreateOrder/EditSlippage/EditSlippageScreen/EditSlippageScreen'
