+++ apps/wallet-mobile/translations/messages/src/TxHistory/TxHistoryNavigator.json
@@ -6,12 +6,12 @@
     "start": {
       "line": 252,
       "column": 16,
-      "index": 8734
+      "index": 8780
     },
     "end": {
       "line": 255,
       "column": 3,
-      "index": 8823
+      "index": 8869
     }
   },
   {
@@ -21,12 +21,12 @@
     "start": {
       "line": 256,
       "column": 13,
-      "index": 8838
+      "index": 8884
     },
     "end": {
       "line": 259,
       "column": 3,
-      "index": 8911
+      "index": 8957
     }
   },
   {
@@ -36,12 +36,12 @@
     "start": {
       "line": 260,
       "column": 17,
-      "index": 8930
+      "index": 8976
     },
     "end": {
       "line": 263,
       "column": 3,
-      "index": 9007
+      "index": 9053
     }
   },
   {
@@ -51,12 +51,12 @@
     "start": {
       "line": 264,
       "column": 15,
-      "index": 9024
+      "index": 9070
     },
     "end": {
       "line": 267,
       "column": 3,
-      "index": 9097
+      "index": 9143
     }
   },
   {
@@ -66,12 +66,12 @@
     "start": {
       "line": 268,
       "column": 21,
-      "index": 9120
+      "index": 9166
     },
     "end": {
       "line": 271,
       "column": 3,
-      "index": 9215
+      "index": 9261
     }
   },
   {
@@ -81,12 +81,12 @@
     "start": {
       "line": 272,
       "column": 14,
-      "index": 9231
+      "index": 9277
     },
     "end": {
       "line": 275,
       "column": 3,
-      "index": 9312
+      "index": 9358
     }
   },
   {
@@ -96,12 +96,12 @@
     "start": {
       "line": 276,
       "column": 13,
-      "index": 9327
+      "index": 9373
     },
     "end": {
       "line": 279,
       "column": 3,
-      "index": 9407
+      "index": 9453
     }
   },
   {
@@ -111,12 +111,12 @@
     "start": {
       "line": 280,
       "column": 18,
-      "index": 9427
+      "index": 9473
     },
     "end": {
       "line": 283,
       "column": 3,
-      "index": 9528
+      "index": 9574
     }
   },
   {
@@ -126,12 +126,12 @@
     "start": {
       "line": 284,
       "column": 20,
-      "index": 9550
+      "index": 9596
     },
     "end": {
       "line": 287,
       "column": 3,
-      "index": 9639
+      "index": 9685
     }
   },
   {
@@ -141,12 +141,12 @@
     "start": {
       "line": 288,
       "column": 26,
-      "index": 9667
+      "index": 9713
     },
     "end": {
       "line": 291,
       "column": 3,
-      "index": 9771
+      "index": 9817
     }
   },
   {
@@ -156,12 +156,12 @@
     "start": {
       "line": 292,
       "column": 19,
-      "index": 9792
+      "index": 9838
     },
     "end": {
       "line": 295,
       "column": 3,
-      "index": 9885
+      "index": 9931
     }
   },
   {
@@ -171,12 +171,12 @@
     "start": {
       "line": 296,
       "column": 16,
-      "index": 9903
+      "index": 9949
     },
     "end": {
       "line": 299,
       "column": 3,
-      "index": 9989
+      "index": 10035
     }
   },
   {
@@ -186,12 +186,12 @@
     "start": {
       "line": 300,
       "column": 19,
-      "index": 10010
+      "index": 10056
     },
     "end": {
       "line": 306,
       "column": 3,
-      "index": 10248
+      "index": 10294
     }
   },
   {
@@ -201,12 +201,12 @@
     "start": {
       "line": 307,
       "column": 27,
-      "index": 10277
+      "index": 10323
     },
     "end": {
       "line": 310,
       "column": 3,
-      "index": 10370
+      "index": 10416
     }
   }
 ]
